fix: for de-dup builds, build into one of the function folder instead of temporary folder (add integration test for the usecase where Gemfile will be in the root folder)
