+++ tests/integration/testdata/buildcmd/Gemfile
@@ -1,0 +1,3 @@
+source "https://rubygems.org"
+
+gem "aws-record", "~> 2"
