+++ tests/integration/testdata/buildcmd/RubyWithRootGemfile/app.rb
@@ -1,0 +1,6 @@
+require 'json'
+require 'httparty'
+
+def lambda_handler(event:, context:)
+  "Hello World"
+end
